+++ client/src/pages/dashboard/DashboardPageContainer.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useState } from "react";
 
 import UsersRepairs from "./UsersRepairs";
-import FilterRepairsTool from "../../components/RepairList/FilterRepairs/FilterRepairsContainer";
+import FilterMenuContainer from "../../components/RepairList/FilterRepairs/FilterMenuContainer";
 import useGetUserRepairs from "../../hooks/useGetUserRepairs";
 import { RepairDataT } from "../../../types";
 // const testRepairList = [new Repair(), new Repair()];
 
 type Filter = {
   category: string;
-  options: [string];
+  option: string;
 };
 
 export default function DashboardPageContainer(): React.ReactNode {
@@ -20,13 +20,17 @@ export default function DashboardPageContainer(): React.ReactNode {
   const [repairList, setRepairList] = useState<RepairDataT[]>(foundRepairs); //data
 
   const [filteredList, setFilteredList] = useState<RepairDataT[]>(repairList); //optionally filtered data
-  const [appliedFilters, setAppliedFilters] = useState<Filter[]>([]); //optionally filtered data
+  const [appliedFilters, setAppliedFilters] = useState<
+    Map<string, Set<string>>
+  >(new Map()); //optionally filtered data
 
   useEffect(() => {
     getUserRepairs(PAGE_LIMIT, 0);
   }, []);
 
   useEffect(() => {
+    console.log("fetch data");
+
     setRepairList(foundRepairs);
     setFilteredList(foundRepairs);
   }, [foundRepairs]);
@@ -38,12 +42,16 @@ export default function DashboardPageContainer(): React.ReactNode {
   return (
     <div className="flex  min-h-screen">
       <aside className=" w-1/6 bg-slate-600">
-        <FilterRepairsTool
+        <FilterMenuContainer
           setFilters={(filter: Filter) => {
-            console.log("set applied filters", filter);
-
             //todo get filters working
-            setAppliedFilters([filter]);
+
+            setAppliedFilters((currentFilters) => {
+              return updateFilter({
+                filters: currentFilters,
+                newFilter: filter,
+              });
+            });
           }}
           list={filteredList}
         />
@@ -100,4 +108,31 @@ function PaganationControls({ onPageChange = () => {} }: paginationProps) {
       </section>
     </>
   );
+}
+
+function updateFilter({
+  filters,
+  newFilter,
+}: {
+  filters: Map<string, Set<string>>;
+  newFilter: Filter;
+}) {
+  //create map with set
+  // key is category
+  //value will be set of strings
+  //check for value before adding
+  //remove if already in set
+  //else add into set
+
+  console.log("newFilter", newFilter);
+
+  const category = filters.get(newFilter.category);
+  if (category) {
+    //has filter category check if value is present
+    const values = category.entries();
+  }
+
+  console.log("updated filters", filters);
+
+  return filters;
 }
